Move price badges back into header
